pyproject.toml
[tool.poetry]
name = "cdk_sqs_lambda_dynamodb"
version = "0.1.0"
description = ""
authors = ["mccreemainwoody <metz.clement@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aws-cdk-lib = "2.137.0"
constructs = ">=10.0.0,<11.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=6.2.5, ^8.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
